+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # ActFramework Change Log
 
 **1.8.29**
+* Provide a default HTML page for async controller response #1209
 * Allow it auto refresh when got 409 error during hot-reloading #1207
 * "The response has already been started" error #1208
 * Update FastJSON to 1.2.61 #1205